requirements.txt
# Hugging Face Integration Reqs
datasets
huggingface_hub

# Misc
sentencepiece
tqdm
omegaconf

# Quantization
torchao-nightly==2024.4.2
